requirements.txt
Flask
Flask-SQLAlchemy
Flask-Script
Flask-Migrate
Objects
requests
WTForms
